+++ packages/plugin-packages/mashroom-browser-cache/type-definitions/index.d.ts
@@ -3,9 +3,12 @@ import type {ExpressRequest, ExpressResponse} from '@mashroom/mashroom/type-defi
 
 export interface MashroomCacheControlService {
     /**
-     * Add the Cache-Control header based on the settings and authentication status
+     * Add the Cache-Control header based on the settings and authentication status.
+     * The resourceCanContainSensitiveInformation parameter defines if the resource could contain some sensitive user data
+     * and the caching should be disabled if a user is authenticated.
      */
     addCacheControlHeader(
+        resourceCanContainSensitiveInformation: boolean,
         request: ExpressRequest,
         response: ExpressResponse,
     ): Promise<void>;