+++ packages/plugin-packages/mashroom-error-pages/type-definitions/index.d.ts
@@ -1,0 +1,9 @@
+import type {ExpressMiddleware} from '@mashroom/mashroom/type-definitions';
+
+export type ErrorMapping = {
+    [statusCode: string]: string | undefined | null;
+}
+
+export interface MashroomErrorPagesMiddleware {
+    middleware(): ExpressMiddleware
+}