+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [unreleased]
 
+ * Portal: Disabled the page caching if the user is authenticated because the back button could reveal sensitive information to other users (on the same machine!).
+   See https://owasp.org/www-project-web-security-testing-guide/latest/4-Web_Application_Security_Testing/04-Authentication_Testing/06-Testing_for_Browser_Cache_Weaknesses
+
 ## 1.6.1 (November 26, 2020)
 
  * Portal: Fixed the problem that incomplete pages and sites could be saved in the Portal (which could lead to a corrupt storage)